layout cleanup
