+++ client/src/props/selection/index.js
@@ -2,7 +2,6 @@ import { useState } from "react";
 import Restaurant from "../restaurant";
 
 export default function Selection(props) {
-  const selection = props.selection;
 
   const [openModal, setOpenModal] = useState(false);
 
@@ -14,28 +13,32 @@ export default function Selection(props) {
   }
 
   return (
-    <section id="cuisine-card-container">
-      {selection.map((sel) => (
-        <div
-          className="cuisine-card"
-          key={sel._id}
-          style={{ cursor: "pointer" }}
-          onClick={() => handleClick(sel)}
-        >
-          <div className="cuisine-card-img">
-            <img src={sel.img} alt={sel.alt} width="100" />
-          </div>
-          <div className="cuisine-card-text">
-            <h4>{sel.city}</h4>
-            <p>{sel.cuisine}</p>
-            <h3>{sel.name}</h3>
-          </div>
-        </div>
-      ))}
 
-      {openModal && (
-        <Restaurant data={restaurantData} closeModal={setOpenModal} />
-      )}
-    </section>
+    <div className="cuisine-gallery">
+      <h4 className="cuisine-title">{props.title}</h4>
+      <section id="cuisine-card-container">
+        {props.data.map((sel) => (
+          <div
+            className="cuisine-card"
+            key={sel._id}
+            style={{ cursor: "pointer" }}
+            onClick={() => handleClick(sel)}
+          >
+            <div className="cuisine-card-img">
+              <img src={sel.img} alt={sel.alt} width="100" />
+            </div>
+            <div className="cuisine-card-text">
+              <h4>{sel.city}</h4>
+              <p>{sel.cuisine}</p>
+              <h3>{sel.name}</h3>
+            </div>
+          </div>
+        ))}
+
+        {openModal && (
+          <Restaurant data={restaurantData} closeModal={setOpenModal} />
+        )}
+      </section>
+    </div>
   );
 }
